+++ spark-confluence/src/main/java/com/k15t/spark/confluence/ConfluenceIframeAdminAppAction.java
@@ -1,8 +1,13 @@
 package com.k15t.spark.confluence;
 
 import com.atlassian.confluence.core.ConfluenceActionSupport;
+import com.atlassian.confluence.setup.BuildInformation;
+import com.atlassian.confluence.user.ConfluenceUser;
+import com.atlassian.user.User;
 import com.k15t.spark.base.util.DocumentOutputUtil;
 
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.List;
 
 
@@ -21,9 +26,39 @@ public abstract class ConfluenceIframeAdminAppAction extends ConfluenceActionSup
     private String body;
 
 
+    private static final Method isConfluenceAdministrator;
+
+    static {
+        if (Integer.parseInt(BuildInformation.INSTANCE.getMarketplaceBuildNumber()) < 20452 /* Confluence 8.8 */) {
+            //noinspection RedundantSuppression
+            try {
+                //noinspection JavaReflectionMemberAccess
+                isConfluenceAdministrator = com.atlassian.confluence.security.PermissionManager.class.getMethod("isConfluenceAdministrator",
+                        User.class);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+        } else {
+            //noinspection RedundantSuppression
+            try {
+                //noinspection JavaReflectionMemberAccess
+                isConfluenceAdministrator = com.atlassian.confluence.security.PermissionManager.class.getMethod("isConfluenceAdministrator",
+                        ConfluenceUser.class);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+        }
+    }
+
+
     @Override
     public boolean isPermitted() {
-        return permissionManager.isConfluenceAdministrator(getAuthenticatedUser());
+        try {
+            // After dropping support for Confluence 8.8, replace this by just calling PermissionManager.isConfluenceAdministrator directly.
+            return (boolean) isConfluenceAdministrator.invoke(permissionManager, getAuthenticatedUser());
+        } catch (RuntimeException | IllegalAccessException | InvocationTargetException e) {
+            throw new RuntimeException("Could not invoke PermissionManager.isConfluenceAdministrator.", e);
+        }
     }
 
 
@@ -64,7 +99,7 @@ public abstract class ConfluenceIframeAdminAppAction extends ConfluenceActionSup
 
     @Override
     public String getSpaQueryString() {
-        return getCurrentRequest().getQueryString();
+        return ConfluenceIframeSparkActionHelper.getQueryString(this);
     }