+++ spark-jira/src/main/java/com/k15t/spark/jira/JiraIframeAppAction.java
@@ -4,6 +4,10 @@ import com.atlassian.jira.web.action.JiraWebActionSupport;
 import com.k15t.spark.atlassian.AtlassianSparkIframeAction;
 import com.k15t.spark.base.util.DocumentOutputUtil;
 
+import jakarta.servlet.http.HttpServletRequest;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.Collections;
 import java.util.List;
 
@@ -33,7 +37,25 @@ public abstract class JiraIframeAppAction extends JiraWebActionSupport implement
 
     @Override
     public String getSpaQueryString() {
-        return getHttpRequest().getQueryString();
+        return getQueryString();
+    }
+
+
+    // Replace this method with a direct call to getHttpRequest().getQueryString() when dropping support for Jira 10 and older.
+    private String getQueryString() {
+        try {
+            Method getHttpRequest = JiraWebActionSupport.class.getMethod("getHttpRequest");
+            Object requestObject = getHttpRequest.invoke(this);
+            if (requestObject instanceof HttpServletRequest) {
+                return ((HttpServletRequest) requestObject).getQueryString();
+            } else if (requestObject != null) {
+                Method getQueryString = requestObject.getClass().getMethod("getQueryString");
+                return (String) getQueryString.invoke(requestObject);
+            }
+        } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
+            throw new RuntimeException("Could not retrieve current request or its query string.", e);
+        }
+        return null;
     }