+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Updated docs and code snippets for `staticParams` interceptor
     * Confluence 8.0 introduced `staticParams` as a replacement for `static-params`, the latter of which has been removed in Confluence 9.0. 
 * Updated host product dependencies to Confluence 8.0.0 and Jira 8.13.0
+* Added compatibility for platform 8 versions of host products and Jakarta Servlet APIs, while still compiling against the older versions in general.
+    * This enables Spark to run on multiple platform versions, such as Confluence versions 8-10.
+    * Spark servlets must therefore implement the Jakarta Servlet API (instead of Javax).
 
 # Spark 3.1.0
 * Introduced support for dark mode, by running `SPARK.initializeTheming()` in the parent window.